+++ crates/sncast/src/helpers/devnet_detection.rs
