+++ crates/sncast/src/response/explorer_link.rs
@@ -1,6 +1,6 @@
 use crate::Network;
 use crate::helpers::{
-    block_explorer::LinkProvider, configuration::CastConfig, devnet,
+    block_explorer::LinkProvider, configuration::CastConfig, devnet_detection,
 };
 use foundry_ui::Message;
 use serde::Serialize;
@@ -70,7 +70,7 @@ where
     let network = chain_id.try_into().ok()?;
 
     let is_devnet_network = matches!(network, Network::Devnet);
-    let is_devnet_running = devnet::is_devnet_running();
+    let is_devnet_running = devnet_detection::is_devnet_running();
     let is_devnet = is_devnet_network || is_devnet_running;
 
     if (!config.show_explorer_links || is_devnet) && !is_explorer_link_overridden() {
